got all code to run properly, needs tests

+++ 19_Capstone/java/src/main/java/com/techelevator/ProductAbstract.java
@@ -1,10 +1,12 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public abstract class ProductAbstract {
 	private String type;
 	private String name;
 	private String code;
-	private Double cost;
+	private BigDecimal cost;
 	private Integer quantity;
 	private String dispenseMessage;
 
@@ -12,7 +14,7 @@ public abstract class ProductAbstract {
 		return dispenseMessage;
 	}
 	
-	public ProductAbstract(String type, String name, String code, Double cost, Integer quantity) {
+	public ProductAbstract(String type, String name, String code, BigDecimal cost, Integer quantity) {
 		this.type = type;
 		this.name = name;
 		this.code = code;
@@ -37,10 +39,10 @@ public abstract class ProductAbstract {
 	public void setCode(String code) {
 		this.code = code;
 	}
-	public Double getCost() {
+	public BigDecimal getCost() {
 		return cost;
 	}
-	public void setCost(Double cost) {
+	public void setCost(BigDecimal cost) {
 		this.cost = cost;
 	}
 	public Integer getQuantity() {
